made signature service standalone
