+++ Backend/ManagerLayer/KFC_SSO_Utility/KFC_SSO_Manager.cs
@@ -5,6 +5,7 @@ using ManagerLayer.AccessControl;
 using ManagerLayer.Logging;
 using ManagerLayer.UserManagement;
 using ManagerLayer.Users;
+using ServiceLayer.KFC_API_Services;
 using ServiceLayer.Services;
 using System;
 using System.Collections.Generic;
@@ -37,9 +38,9 @@ namespace ManagerLayer.KFC_SSO_Utility
             ////////////////////////////////////////
             /// User oAuth at the indivudal application level
             // verify if the login payload is valid via its signature
-            var _ssoServiceAuth = new KFC_SSO_APIService.RequestPayloadAuthentication();
+            var _ssoServiceAuth = new SignatureService();
             loggingManager = new LoggingManager();
-            if (!_ssoServiceAuth.IsValidClientRequest(ssoID, Username, timestamp, signature))
+            if (!_ssoServiceAuth.IsValidClientRequest(ssoID.ToString(), Username, timestamp, signature))
             {
                 newLog = new LogRequestDTO(ssoID.ToString(), Username,
                         "Login/Registration API", Username, "Invalid signing attempt",
@@ -71,7 +72,7 @@ namespace ManagerLayer.KFC_SSO_Utility
 
         public async Task<bool> DeleteUserFromSSOviaPointmap(User user)
         {
-            var _ssoAPI = new KFC_SSO_APIService();
+            var _ssoAPI = new SSO_APIService();
             var requestResponse = await _ssoAPI.DeleteUserFromSSO(user);
             if (requestResponse.IsSuccessStatusCode)
             {
